finish api

+++ RESTfulApiHibernateSercurity/DemooRESTfulAPI/src/main/java/com/xamthien/DAO/FlightSchedulesDAO.java
@@ -27,7 +27,7 @@ Session session;
 //		Query que = session.createQuery(sql);
 //        return que.list();
 //    }
-    public FlightSchedules getFlightSchedulesByID(int id) {
+    public FlightSchedules getFlightSchedulesByID(long id) {
     	FlightSchedules bk = null;
          try {
         	 session = getSession();
